+++ Packs/ServiceNow/Integrations/ServiceNowv2/test_data/result_constants.py
@@ -278,6 +278,11 @@ EXPECTED_GET_TABLE_NAME = {
         }
     ]
 }
+EXPECTED_ADD_TAG = {
+    'ServiceNow.Ticket(val.ID===obj.ID)': {
+        'ID': '123', 'TagTitle': 'title', 'TagID': '1234'
+    }
+}
 EXPECTED_DOCUMENT_ROUTE = {
     'ServiceNow.WorkItem(val.WorkItemID===obj.WorkItemID)':
         {