+++ Packs/ServiceNow/Integrations/ServiceNowv2/CHANGELOG.md
@@ -1,7 +1,7 @@
 ## [Unreleased]
   - Added the ***servicenow-document-route-to-queue*** command.
+  - Added the ***servicenow-add-tag*** command.
   - Added the *system_params* argument to the ***servicenow-query-table***, ***servicenow-query-tickets*** commands.
-
 
 ## [20.4.1] - 2020-04-29
   - Added the *additional_fields* argument to the following commands: